fix the runtime bugs of data_ptr symbol not found and others

+++ scene/gaussian_model.py
@@ -357,7 +357,7 @@ class GaussianModel:
 
         stds = self.get_scaling[selected_pts_mask].repeat(N,1)
         means =torch.zeros((stds.size(0), 3),device="cuda", dtype=torch.half)
-        samples = torch.normal(mean=means, std=stds, dtype=torch.half).half()
+        samples = torch.normal(mean=means, std=stds).half()
         rots = build_rotation(self._rotation[selected_pts_mask]).repeat(N,1,1)
         new_xyz = torch.bmm(rots, samples.unsqueeze(-1)).squeeze(-1) + self.get_xyz[selected_pts_mask].repeat(N, 1)
         new_scaling = self.scaling_inverse_activation(self.get_scaling[selected_pts_mask].repeat(N,1) / (0.8*N))
